# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Zeno API
  version: 1.5.0
  description: >-
    API runtime untuk dashboard dan personal workspace Zeno. Manual Learning:
    regular users access only their own records (foreign IDs return 404); admins
    can list and manage records of every owner, including files and materials.
    New modules belong to their creator; sessions inherit their module owner.
    Editing never transfers ownership or reparents a session across owners.
servers:
  - url: http://localhost:3000
paths:
...
  /api/learning:
    get:
      tags: [Learning]
      summary: Baca learning journal
      parameters:
        - {in: query, name: date, schema: {type: string, format: date}}
      responses:
        '200': {description: Learning entries}
    post:
      tags: [Learning]
      summary: Tambah learning entry
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningInput'}
      responses:
        '201':
          description: Entry dibuat
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningEntry'}
        '400': {description: Input tidak valid}
  /api/learning/{id}:
    put:
      tags: [Learning]
      summary: Edit learning entry atau ubah status checklist
      parameters:
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningInput'}
      responses:
        '200': {description: Entry diperbarui}
        '400': {description: Input tidak valid}
        '404': {description: Entry tidak ditemukan}
    delete:
      tags: [Learning]
      summary: Hapus learning entry
      parameters:
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
        - {in: query, name: date, required: true, schema: {type: string, format: date}}
      responses:
        '200': {description: Entry dihapus}
        '404': {description: Entry tidak ditemukan}
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: zeno_session
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
...
    LearningInput:
      type: object
      required: [date, title]
      properties:
        date: {type: string, format: date}
        title: {type: string, maxLength: 160}
        note: {type: string, maxLength: 2000}
        category: {type: string, maxLength: 60, default: General}
        completed: {type: boolean, default: false}
    LearningEntry:
      allOf:
        - {$ref: '#/components/schemas/LearningInput'}
        - type: object
          properties:
            id: {type: string, format: uuid}
            createdAt: {type: string, format: date-time}
...